% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build.RF.ITR.R
\name{Build.RF.ITR}
\alias{Build.RF.ITR}
\title{Builds the random forest of interaction trees}
\usage{
Build.RF.ITR(dat, col.y, col.trt, col.prtx, split.var, ctg = NA, N0 = 20,
  n0 = 5, max.depth = 10, ntree = 500,
  mtry = max(floor(length(split.var)/3), 1), avoid.nul.tree = F)
}
\arguments{
\item{dat}{the data set being used to grow the random forest. Required input.}

\item{col.y}{the response variable. Required input.}

\item{col.trt}{the treatment indicator.  Must be binary. Required input.}

\item{col.prtx}{the probability of being assigned to treatment group. Required input.}

\item{split.var}{vector of columns containing the desired splitting variables.  Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NA.  Not available yet.}

\item{N0}{minimum number of observations needed to call a node terminal.  Defaults to 20.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{max.depth}{controls the maximum depth of the tree. Defaults to 10.}

\item{ntree}{sets the number of trees to be generated. Defaults to 500.}

\item{mtry}{sets the number of randomly selected splitting variables to be included. Defaults to max of length(split.var)/3 rounded down and 1.}

\item{avoid.nul.tree}{controls if trees with no splits (null trees) are allowed. Defaults to FALSE.}
}
\value{
summary of randomly generated trees (summary done by tree)
}
\description{
Builds the random forest of interaction trees
}
\examples{
forest<-Build.RF.ITR(dat=data, col.y="y", col.trt="trt", col.prtx="prtx", 
                     split.var=3:7)
This builds a forest of 500 trees using the dataset called 'data' with columns
'y', 'trt', and 'prtx' for the outcome, treatement indicator, and probability of being
in treatment group, respectively.  The splitting variables are found in columns 3-7.
}

