\name{dataOBS}
\alias{dataOBS}
\docType{data}
\title{
Simulated data from and observational design with 7 predictors (4 continuous and 3 categorical) on a continuous outcome. 
}
\description{
Simulated data from an RCT design with 7 predictors (4 continous and 3 categorical)
}
\usage{data("dataRCT")}
\format{
  A simulated data frame with 1000 observations on 4 continuous predictors (X1 - X4), three categorical predictors (c1 - c3), a single outcome (y), binary treatment (trt), and associated propensity score (prtx). 
}
\details{
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
