\name{datCAT}
\alias{datCAT}
\docType{data}
\title{
Simulated data from an RCT design with only categorical predictors (4 total).
}
\description{
Simulated data from an RCT design with 4 categorical predictors.
}
\usage{data("dataRCT")}
\format{
  A simulated data frame with 1000 observations on 4 categorical predictors (c1 - c4), a single outcome (y), binary treatment (trt), and associated propensity score (prtx). 
}
\keyword{datasets}
