% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{de}
\alias{de}
\title{Finds all decendants of a particular node in a tree.  Used in the send.down.VI.ITR function.}
\usage{
de(x, tree)
}
\arguments{
\item{x}{node}

\item{tree}{interaction tree}
}
\value{
summary of decendant nodes
}
\description{
Finds all decendants of a particular node in a tree.  Used in the send.down.VI.ITR function.
}

