% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.ITR.R
\name{grow.ITR}
\alias{grow.ITR}
\title{Grows a large interaction tree}
\usage{
grow.ITR(data, test = NULL, min.ndsz = 20, n0 = 5, split.var,
  ctg = NULL, max.depth = 15, mtry = length(split.var))
}
\arguments{
\item{data}{data set from which the tree is to be grown.  Must contain outcome, binary 
treatment indicator, columns of splitting covariates, and column of probability of being
in treatment group.}

\item{min.ndsz}{minimum number of observations required to call a node terminal. Defaults to 20.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{split.var}{columns of potential spliting variables. Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NULL.  Not available yet.}

\item{max.depth}{controls the maximum depth of the tree. Defaults to 15.}

\item{mtry}{sets the number of randomly selected splitting variables to be included. Defaults to number of splitting variables.}

\item{N0}{minimum number of observations needed to call a node terminal.  Defaults to 20.}
}
\value{
summary of single interaction tree
}
\description{
Grows a large interaction tree
}
\examples{
tree<-grow.ITR(data=rctdata, split.var=3:7)
Generates tree using rctdata with potential splitting variables located in columns 3-7.
}

