% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrtest2.R
\name{itrtest2}
\alias{itrtest2}
\title{Value function used for initial treatment heterogeneity assessment.  This is used 
inside the tree growing function.}
\usage{
itrtest2(dat, z, n0)
}
\arguments{
\item{dat}{dataset being assessed}

\item{z}{new (alternative) treatment assignment in the splitting procedure.}

\item{n0}{minimum number of observations needed to make a split.}
}
\value{
itr value from a defined split.
}
\description{
Value function used for initial treatment heterogeneity assessment.  This is used 
inside the tree growing function.
}

