% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VI.R
\name{plot.VI}
\alias{plot.VI}
\title{Plots the variable importance measures. Requires an input of a variable importance object
created from the function Variable.Importance.ITR.}
\usage{
\method{plot}{VI}(VI, filename = NULL, horizontal = T, rain.bow = T)
}
\arguments{
\item{VI}{variable importance object from Variable.Importance.ITR. Required input.}

\item{filename}{name of destination file. Defaults to NULL.}

\item{horizontal}{barplot horizontal? Defaults to TRUE.}

\item{rain.bow}{controls color scheme for output graphic. Defaults to TRUE.}
}
\value{
barplot of variable importance measures
}
\description{
Plots the variable importance measures. Requires an input of a variable importance object
created from the function Variable.Importance.ITR.
}

