% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.set.R
\name{power.set}
\alias{power.set}
\title{Categorical predictor subsetting in the tree growing function.}
\usage{
power.set(x)
}
\arguments{
\item{x}{a categorical predictor.}
}
\value{
Returns a list of the subsets of the predictor to be considered.
}
