% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ITR.R
\name{predict.ITR}
\alias{predict.ITR}
\title{Treatment Prediction Function}
\usage{
\method{predict}{ITR}(input, new.dat)
}
\arguments{
\item{input}{tree or forest object from `grow.ITR` or `Build.RF.ITR`.}

\item{new.dat}{data for which predictions are desired}
}
\description{
Used to make treatment prediction for a single tree or random forest
}
\examples{
dat <- gdataM(n=1000, depth=2, beta1=3, beta2=1)
#Build a forest with 100 trees
forest <- Build.RF.ITR(dat, col.y="y", col.trt="trt", col.prtx="prtx", split.var=1:4, ntree=100)
#Predict treatment assignments for 1000 observations in `dat`
predict.ITR(forest, dat)
}
