% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prunes a tree for a given penalty value}
\usage{
prune(tre, a, train, test = NULL, AIPWE = F)
}
\arguments{
\item{tre}{sets the tree to be pruned}

\item{a}{sets the value of the splitting penalty}

\item{train}{the training data used to create the tree}

\item{test}{the testing data to be used.  Defaults to NULL.}

\item{AIPWE}{indicator for AIPWE estimation.}
}
\value{
summary of pruned branches and the associated value of the tree after pruning
}
\description{
The `prune` function allows the user to specify a value of the penalty for a given tree. 
This function uses the "weakest link" criteria in order to evaluate the order in which branches are pruned
and gives the penalized value along with the unpenalized value. If testing data are provided for validation, 
then the penalized and unpenalized values from the testing data run down the tree structure are also provided.
}
\examples{

}
