% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PruneCV.R
\name{PruneCV}
\alias{PruneCV}
\title{This function is outdated and it is recommended that users use treeCV instead.
Performs k-fold cross validation for selection optimal tuning parameter lambda.}
\usage{
PruneCV(data, param, min.ndsz = 20, n0 = 5, split.var, nfolds = 10,
  graphic = FALSE, graphicname = NULL)
}
\arguments{
\item{data}{input data}

\item{param}{vector of pruning parameter values to be considered}

\item{min.ndsz}{sets the minimal number of observations allowed in terminal nodes. Defaults to 20.}

\item{n0}{sets the minimum number of observations from each treatment group to be in each child node.  Defaults to 5.}

\item{split.var}{specifies the columns of splitting variables in the input data.}

\item{nfolds}{defaults to 10.}

\item{graphic}{logical indicator for whether CV graphic should be generated.  Defaults to FALSE.}

\item{graphicname}{filename of the CV graphic.  Must be specified if graphic=TRUE.  Defaults to FALSE.}
}
\value{
summary of pruned branches and the associated value of the tree after pruning
}
\description{
This function is outdated and it is recommended that users use treeCV instead.
Performs k-fold cross validation for selection optimal tuning parameter lambda.
}
\examples{

}
