% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdat.R
\name{rdat}
\alias{rdat}
\title{Simulates data from an RCT according to the following model:
2 + 2*sign(x1<cut2) + beta1*trt*subgrp + beta2*(1-trt)*subgrp + N(0,sigma)
If depth=1, then subgrp=(x1<=cut1)
If depth!=1 then subgrp=(x1>=0.3 & x3>=0.1)}
\usage{
rdat(n = 100, K = 50, beta1 = 2, beta2 = 2, sigma = 1, cut1 = 0.5,
  cut2 = 0.5, depth = 1)
}
\arguments{
\item{n}{size of the dataset to be generate.  Defaults to 100.}

\item{beta1}{controls the strength of the treatment effect. Defaults to 2.}

\item{beta2}{controls the strength of the noise. Defaults to 2.}

\item{sigma}{controls standard deviation of random variation.  Defaults to 1.}

\item{depth}{gives the number of interacting covariates. If set to 1, then 
then covariate X1 interacts with treatment. If set to another value, then 
covariates X1 and X3 both interact with treatment effect (one-way interactions). Defaults to 1.}

\item{cut1/cut2}{controls where the cutpoints are to define subgroups.}
}
\value{
dataframe containing y (outcome), x1-x4 (covariates), trt (treatment), prtx (probability of being in treatment group)
}
\description{
Simulates data from an RCT according to the following model:
2 + 2*sign(x1<cut2) + beta1*trt*subgrp + beta2*(1-trt)*subgrp + N(0,sigma)
If depth=1, then subgrp=(x1<=cut1)
If depth!=1 then subgrp=(x1>=0.3 & x3>=0.1)
}
\examples{
data<-rdat(n=500)
This generates a dataframe with 500 observations, X1 as the only variable interacting with 
the treatment, and a signal to noise ratio of 2/2=1.
}

