% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.VI.ITR.R
\name{send.down.VI.ITR}
\alias{send.down.VI.ITR}
\title{Sends testing data down a tree to assess the performance of the tree. This is used inside
the variable importance function.}
\usage{
send.down.VI.ITR(dat.new, tre, col.y, col.trt, col.prtx, ctg = NA, n0 = 5,
  revise.tree = T, depth = 1)
}
\arguments{
\item{dat.new}{the new data set being sent down the tree. Required input.}

\item{col.y}{the response variable. Required input.}

\item{col.trt}{the treatment indicator.  Must be binary. Required input.}

\item{col.prtx}{the probability of being assigned to treatment group. Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NA.  Not available yet.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{tree}{constructed tree.}
}
\value{
summary of tree performance
}
\description{
Sends testing data down a tree to assess the performance of the tree. This is used inside
the variable importance function.
}

