% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable.Importance.ITR.R
\name{Variable.Importance.ITR}
\alias{Variable.Importance.ITR}
\title{Calcuates variable importance measures for a random forest object.  Input must be an object
 from the random forest function Build.RF.ITR.}
\usage{
Variable.Importance.ITR(RF.fit, n0 = 2, sort = T, details = F,
  truncate.zeros = T, depth = 1)
}
\arguments{
\item{RF.fit}{forest object from Build.RF.ITR. Required input.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 2.}

\item{sort}{sort the variable importance measure? Defaults to TRUE.}

\item{details}{print details. Defaults to FALSE.}

\item{truncate.zeros}{sets variable importances less than 0 to 0. Defaults to TRUE.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}
}
\value{
summary of tree performance
}
\description{
Calcuates variable importance measures for a random forest object.  Input must be an object
 from the random forest function Build.RF.ITR.
}

