% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prunes a bootstrap tree with penalty (a) for additional splitting}
\usage{
prune(tre, a, train, test = NULL, AIPWE = F)
}
\arguments{
\item{tre}{sets the tree to be pruned}

\item{a}{sets the value of the splitting penalty}

\item{train}{the training data used to create the tree}

\item{test}{the testing data to be used.  Defaults to NULL.}

\item{AIPWE}{indicator for AIPWE estimation.}
}
\value{
summary of pruned branches and the associated value of the tree after pruning
}
\description{
Prunes a bootstrap tree with penalty (a) for additional splitting
}
\examples{

}
