% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.R
\name{send.down}
\alias{send.down}
\title{Sends testing data down a tree to assess the performance of the tree. This is used inside
the variable importance function.}
\usage{
send.down(dat.new, tre, char.var = 1000, ctgs = NULL)
}
\arguments{
\item{tre}{tree object from grow.ITR().  Required input.}

\item{ctgs}{categorical variables, entered as columns in `dat.new`}

\item{dat}{data to be run down the tree.  Required input.}
}
\value{
summary of node assignments for `dat.new` run down tree `tre`
}
\description{
Sends testing data down a tree to assess the performance of the tree. This is used inside
the variable importance function.
}
