% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.VI.ITR.R
\name{send.down.VI.ITR}
\alias{send.down.VI.ITR}
\title{Sends testing data and permuted testing data down a bootstrap tree from a forest object.}
\usage{
send.down.VI.ITR(dat.new, tre, col.y, col.trt, col.prtx, ctg = NULL,
  n0 = n0, N0 = N0, revise.tree = T, depth = 1, AIPWE = AIPWE)
}
\arguments{
\item{dat.new}{the new data set being sent down the tree. Required input.}

\item{col.y}{the response variable. Required input.}

\item{col.trt}{the treatment indicator.  Must be binary. Required input.}

\item{col.prtx}{the probability of being assigned to treatment group. Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NA.  Not available yet.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{AIPWE}{indicator for AIPWE estimation.}

\item{tree}{constructed tree.}
}
\value{
summary of tree performance
}
